+++ hadoop-hdfs-project/hadoop-hdfs-client/src/main/java/org/apache/hadoop/hdfs/client/HdfsClientConfigKeys.java
@@ -19,8 +19,15 @@ package org.apache.hadoop.hdfs.client;
 
 /** Client configuration properties */
 public interface HdfsClientConfigKeys {
-  static final String PREFIX = "dfs.client.";
+  String  DFS_BLOCK_SIZE_KEY = "dfs.blocksize";
+  long    DFS_BLOCK_SIZE_DEFAULT = 128*1024*1024;
+  String  DFS_REPLICATION_KEY = "dfs.replication";
+  short   DFS_REPLICATION_DEFAULT = 3;
+  String  DFS_WEBHDFS_USER_PATTERN_DEFAULT = "^[A-Za-z_][A-Za-z0-9._-]*[$]?$";
+  String DFS_WEBHDFS_ACL_PERMISSION_PATTERN_DEFAULT =
+      "^(default:)?(user|group|mask|other):[[A-Za-z_][A-Za-z0-9._-]]*:([rwx-]{3})?(,(default:)?(user|group|mask|other):[[A-Za-z_][A-Za-z0-9._-]]*:([rwx-]{3})?)*$";
 
+  static final String PREFIX = "dfs.client.";
   /** Client retry configuration properties */
   public interface Retry {
     static final String PREFIX = HdfsClientConfigKeys.PREFIX + "retry.";
