+++ hadoop-hdfs-project/hadoop-hdfs-client/src/main/java/org/apache/hadoop/hdfs/web/resources/EnumSetParam.java
@@ -63,12 +63,9 @@ abstract class EnumSetParam<E extends Enum<E>>
   public String getValueString() {
     return toString(value);
   }
-  
-  /**
-   * The domain of the parameter.
-   */
-  static final class Domain<E extends Enum<E>>
-      extends Param.Domain<EnumSet<E>> {
+
+  /** The domain of the parameter. */
+  static final class Domain<E extends Enum<E>> extends Param.Domain<EnumSet<E>> {
     private final Class<E> enumClass;
 
     Domain(String name, final Class<E> enumClass) {
