+++ hadoop-hdfs-project/hadoop-hdfs-client/src/main/java/org/apache/hadoop/hdfs/web/resources/PermissionParam.java
@@ -42,7 +42,7 @@ public class PermissionParam extends ShortParam {
   public static FsPermission getDefaultFsPermission() {
     return new FsPermission(DEFAULT_PERMISSION);
   }
-  
+
   /**
    * Constructor.
    *
